+++ CHANGELOG.md
@@ -54,6 +54,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
 
+
+## v2.44.1 - 2026-03-02
+
+### Added
+- added "Reset All Movers" button to profiles tab
+
+### Changed
+- no cursor-anchoring for tooltips in combat anymore to avoid taints
+
+### Fixed
+- re-apply frame anchors after profile change
+- minor objective tracking skinning fixes
+- prevent CDM flash on load
+- fixed ESC and slash commands not working in Edit Mode
 ## v2.44.0 - 2026-03-02
 
 ### Added